Small adjustments for native Mac bundle.
Works fine, but porting all the helpers to work from the bundle would be a
lot of work. All must be changed to find their resources inside the .app
Proof of concept done with antiword. Does not make sense to ship without
at least poppler (and maybe unrtf, mutagen, exiftool).