GUI reslist para format: define %s for explicit snippets abstract.
If %s is requested, %A only outputs the document own abstract.
Add %s to the default paragraph format.
This allows using "abstract" for the finder comments on the Mac, because it complements the snippets
instead of replacing them.
Misc code cleanups in reslistpager.cpp (use strings instead of streams, factorize stuff, etc.)